lsp-record: A utility to record and replay LSP requests (#37)
This is an initial version of `lsp-record`. It both records and replays LSP
sessions, supports our remote `git://` style root URIs, and automated parts of
docker build/run. This allows us to record a session using Sourcegraph, and
then later play it back to confirm if the output still matches. Still to come
is the infrastructure around confirming if the recorded outputs still match is
missing, but the bulk of the work is done.