lsp-record: Allow specifying a testcase name (#38)
To simplify saving the testdata we allow a testcase name to be passed as an
argument. Instead of stdin and stdout, a .input.json and/or .golden.json file
is used instead.

Right now lsp-record test will always write to the files, and not actually
"test" by comparing to the golden output. For now this is fine since we expect
the files to be checked in, and can be added as a follow-up PR.